environments/production/source/requirements.txt
pygithub==1.59.1
cryptography==3.4.7
urllib3==1.26.6 
semantic-version==2.10.0
